refactor: Extract file upload logic from safe loading component
